+++ .anvil_editor.yaml
@@ -251,7 +251,6 @@ unique_ids:
     admin.dashboard.loan_management.application_intake.Approved_loans: '1719377476665565150189967.4034'
     admin.dashboard.manage_ascend_score: '170953922392368996895390.69653'
     admin.dashboard.loan_management.application_intake.Rejected_loans.RowTemplate65: '1719380937839710685551952.6975'
-    borrower.borrower_registration_forms.borrower_registration_form_nav_bar: '1720671593380515479005239.986'
     borrower.dashboard.extension_loan_request: SOVMQQNDKL4DAN7TINAJVUZANBR4RWQY
     admin.dashboard.manage_cms.add_borrower_dropdown_details.edit_qualification: '1706191359421651665929783.5277'
     bank_users.main_form.signin_page.forgot_password: '1720411186028296089758662.0104'
