--- client_code/borrower/borrower_registration_forms/borrower_registration_form_nav_bar/__init__.py
@@ -1,29 +1,0 @@
-from ._anvil_designer import borrower_registration_form_nav_barTemplate
-from anvil import *
-import anvil.server
-import anvil.google.auth, anvil.google.drive
-from anvil.google.drive import app_files
-import anvil.users
-import anvil.tables as tables
-import anvil.tables.query as q
-from anvil.tables import app_tables
-
-
-class borrower_registration_form_nav_bar(borrower_registration_form_nav_barTemplate):
-    def __init__(self, **properties):
-        # Set Form properties and Data Bindings.
-        self.init_components(**properties)
-
-        # Any code you write here will run before the form opens.
-
-    def home_main_form_link_click(self, **event_args):
-        """This method is called when the link is clicked"""
-        open_form('bank_users.main_form')
-
-    def about_main_form_link_click(self, **event_args):
-        """This method is called when the link is clicked"""
-        open_form('bank_users.main_form.about_main_form')
-
-    def contact_main_form_link_click(self, **event_args):
-        """This method is called when the link is clicked"""
-        open_form('bank_users.main_form.contact_main_form')
