+++ client_code/borrower/borrower_registration_forms/borrwer_registration_navigation_bar/__init__.py
@@ -14,20 +14,16 @@ class borrwer_registration_navigation_bar(borrwer_registration_navigation_barTem
     # Set Form properties and Data Bindings.
     self.init_components(**properties)
 
-    # Any code you write here will run before the form opens.
+    
+  def home_main_form_link_click(self, **event_args):
+        """This method is called when the link is clicked"""
+        open_form('bank_users.main_form')
 
-  def borrower_dashboard_home_linkhome_borrower_registration_button_copy_1_click(self, **event_args):
-    """This method is called when the link is clicked"""
-    open_form('bank_users.main_form.investNow_applyForLoan')
+  def about_main_form_link_click(self, **event_args):
+        """This method is called when the link is clicked"""
+        open_form('bank_users.main_form.about_main_form')
 
   def contact_main_form_link_click(self, **event_args):
-    """This method is called when the link is clicked"""
-    open_form('borrower.borrower_registration_forms.dashboard_contact_copy')
-
-  def About_Us_click(self, **event_args):
-    """This method is called when the link is clicked"""
-    open_form('borrower.borrower_registration_forms.dashboard_about_borrower_registration_form')
-
-  def Report_A_Problem_click(self, **event_args):
-    """This method is called when the link is clicked"""
-    open_form('borrower.borrower_registration_forms.dashboard_report_a_problem_copy')
+        """This method is called when the link is clicked"""
+        open_form('bank_users.main_form.contact_main_form')  
+
